refactor(label): remove obsolete Label data class and related test
Deleted the `Label` data class from `renderables.label` as it was replaced by the new `Label` class with OpenGL functionality.
