+++ app/src/androidTest/java/com/github/lookupgroup27/lookup/util/opengl/LabelUtilsTest.kt
@@ -1,4 +1,4 @@
-package com.github.lookupgroup27.lookup.model.map.renderables
+package com.github.lookupgroup27.lookup.util.opengl
 
 import android.graphics.Bitmap
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -8,20 +8,7 @@ import org.junit.runner.RunWith
 
 /** Instrumented tests for LabelUtils and the Label class. */
 @RunWith(AndroidJUnit4::class)
-class LabelTest {
-
-  /** Tests that a Label object can be created successfully. */
-  @Test
-  fun testLabelCreation() {
-    val text = "Star A"
-    val position = floatArrayOf(1.0f, 2.0f, 3.0f)
-    val label = Label(text, position)
-
-    assertEquals("Label text should be 'Star A'", "Star A", label.text)
-    assertArrayEquals(
-        "Label position should match", floatArrayOf(1.0f, 2.0f, 3.0f), label.position, 0.0f)
-    assertNull("Label textureId should be null by default", label.textureId)
-  }
+class LabelUtilsTest {
 
   /** Tests that the bitmap is created successfully. */
   @Test
